@@ -1,7 +1,7 @@
-import { AnyType } from "../types"
-import { CoreValidator, TreeTraverser } from "../core/validator"
-import { BaseValidator } from "../base/validator"
-import { RequiredValidator } from "../required/validator"
+import { AnyType } from "../types.js"
+import { CoreValidator, TreeTraverser } from "../core/validator.js"
+import { BaseValidator } from "../base/validator.js"
+import { RequiredValidator } from "../required/validator.js"
 
 
 export class AnyOfValidator< T > extends BaseValidator< T, AnyOfValidator< T > >
